+++ ecd-20200930/src/Models/DescribeDesktopMetadataResponseBody/desktops.php
@@ -1,0 +1,261 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Ecd\V20200930\Models\DescribeDesktopMetadataResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class desktops extends Model
+{
+    /**
+     * @var string
+     */
+    public $chargeType;
+
+    /**
+     * @var string
+     */
+    public $creationTime;
+
+    /**
+     * @var string
+     */
+    public $desktopGroupId;
+
+    /**
+     * @var string
+     */
+    public $desktopId;
+
+    /**
+     * @var string
+     */
+    public $desktopName;
+
+    /**
+     * @var string
+     */
+    public $desktopStatus;
+
+    /**
+     * @var string
+     */
+    public $expiredTime;
+
+    /**
+     * @var string
+     */
+    public $imageId;
+
+    /**
+     * @var string
+     */
+    public $localName;
+
+    /**
+     * @var string[]
+     */
+    public $managementFlags;
+
+    /**
+     * @var string
+     */
+    public $officeSiteId;
+
+    /**
+     * @var string
+     */
+    public $regionId;
+
+    /**
+     * @var string
+     */
+    public $resourceGroupId;
+
+    /**
+     * @var string
+     */
+    public $resourceGroupName;
+
+    /**
+     * @var string
+     */
+    public $startTime;
+    protected $_name = [
+        'chargeType' => 'ChargeType',
+        'creationTime' => 'CreationTime',
+        'desktopGroupId' => 'DesktopGroupId',
+        'desktopId' => 'DesktopId',
+        'desktopName' => 'DesktopName',
+        'desktopStatus' => 'DesktopStatus',
+        'expiredTime' => 'ExpiredTime',
+        'imageId' => 'ImageId',
+        'localName' => 'LocalName',
+        'managementFlags' => 'ManagementFlags',
+        'officeSiteId' => 'OfficeSiteId',
+        'regionId' => 'RegionId',
+        'resourceGroupId' => 'ResourceGroupId',
+        'resourceGroupName' => 'ResourceGroupName',
+        'startTime' => 'StartTime',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->managementFlags)) {
+            Model::validateArray($this->managementFlags);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->chargeType) {
+            $res['ChargeType'] = $this->chargeType;
+        }
+
+        if (null !== $this->creationTime) {
+            $res['CreationTime'] = $this->creationTime;
+        }
+
+        if (null !== $this->desktopGroupId) {
+            $res['DesktopGroupId'] = $this->desktopGroupId;
+        }
+
+        if (null !== $this->desktopId) {
+            $res['DesktopId'] = $this->desktopId;
+        }
+
+        if (null !== $this->desktopName) {
+            $res['DesktopName'] = $this->desktopName;
+        }
+
+        if (null !== $this->desktopStatus) {
+            $res['DesktopStatus'] = $this->desktopStatus;
+        }
+
+        if (null !== $this->expiredTime) {
+            $res['ExpiredTime'] = $this->expiredTime;
+        }
+
+        if (null !== $this->imageId) {
+            $res['ImageId'] = $this->imageId;
+        }
+
+        if (null !== $this->localName) {
+            $res['LocalName'] = $this->localName;
+        }
+
+        if (null !== $this->managementFlags) {
+            if (\is_array($this->managementFlags)) {
+                $res['ManagementFlags'] = [];
+                $n1 = 0;
+                foreach ($this->managementFlags as $item1) {
+                    $res['ManagementFlags'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->officeSiteId) {
+            $res['OfficeSiteId'] = $this->officeSiteId;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->resourceGroupId) {
+            $res['ResourceGroupId'] = $this->resourceGroupId;
+        }
+
+        if (null !== $this->resourceGroupName) {
+            $res['ResourceGroupName'] = $this->resourceGroupName;
+        }
+
+        if (null !== $this->startTime) {
+            $res['StartTime'] = $this->startTime;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ChargeType'])) {
+            $model->chargeType = $map['ChargeType'];
+        }
+
+        if (isset($map['CreationTime'])) {
+            $model->creationTime = $map['CreationTime'];
+        }
+
+        if (isset($map['DesktopGroupId'])) {
+            $model->desktopGroupId = $map['DesktopGroupId'];
+        }
+
+        if (isset($map['DesktopId'])) {
+            $model->desktopId = $map['DesktopId'];
+        }
+
+        if (isset($map['DesktopName'])) {
+            $model->desktopName = $map['DesktopName'];
+        }
+
+        if (isset($map['DesktopStatus'])) {
+            $model->desktopStatus = $map['DesktopStatus'];
+        }
+
+        if (isset($map['ExpiredTime'])) {
+            $model->expiredTime = $map['ExpiredTime'];
+        }
+
+        if (isset($map['ImageId'])) {
+            $model->imageId = $map['ImageId'];
+        }
+
+        if (isset($map['LocalName'])) {
+            $model->localName = $map['LocalName'];
+        }
+
+        if (isset($map['ManagementFlags'])) {
+            if (!empty($map['ManagementFlags'])) {
+                $model->managementFlags = [];
+                $n1 = 0;
+                foreach ($map['ManagementFlags'] as $item1) {
+                    $model->managementFlags[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['OfficeSiteId'])) {
+            $model->officeSiteId = $map['OfficeSiteId'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['ResourceGroupId'])) {
+            $model->resourceGroupId = $map['ResourceGroupId'];
+        }
+
+        if (isset($map['ResourceGroupName'])) {
+            $model->resourceGroupName = $map['ResourceGroupName'];
+        }
+
+        if (isset($map['StartTime'])) {
+            $model->startTime = $map['StartTime'];
+        }
+
+        return $model;
+    }
+}